requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
httpx==0.24.1
aiofiles==24.1.0
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
numpy==1.24.3
tensorflow-macos==2.9.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow==2.9.0; sys_platform != 'darwin' or platform_machine != 'arm64'
opencv-python-headless==4.8.0.74
scikit-learn==0.24.2
scipy==1.7.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
torch==2.0.1
torchvision==0.15.2
geopy==2.2.0
pydantic==1.8.2
